chore(frontend): rename branding asset to report-pilot.png
